moves the last interfaces and put the in packages.
