+++ service-api/src/main/java/org/orbisgis/orbiswps/serviceapi/operations/WPS_2_0_Operations.java
@@ -37,7 +37,7 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.orbiswps.serviceapi;
+package org.orbisgis.orbiswps.serviceapi.operations;
 
 import net.opengis.wps._2_0.*;
