+++ service-api/src/main/java/org/orbisgis/orbiswps/serviceapi/process/ProcessMetadata.java
@@ -37,7 +37,7 @@
  * or contact directly:
  * info_at_ orbisgis.org
  */
-package org.orbisgis.orbiswps.serviceapi;
+package org.orbisgis.orbiswps.serviceapi.process;
 
 /**
  * This class contains the process metadata used by OrbisGIS. A process metadata is composed of a role and a title.
