+++ service/src/main/java/org/orbisgis/orbiswps/service/WpsServerImpl.java
@@ -41,7 +41,9 @@ package org.orbisgis.orbiswps.service;
 
 import net.opengis.ows._2.*;
 import net.opengis.wps._2_0.*;
-import org.orbisgis.orbiswps.serviceapi.ProcessIdentifier;
+import org.orbisgis.orbiswps.serviceapi.operations.WPS_1_0_0_Operations;
+import org.orbisgis.orbiswps.serviceapi.operations.WPS_2_0_Operations;
+import org.orbisgis.orbiswps.serviceapi.process.ProcessIdentifier;
 import org.orbisgis.orbiswps.serviceapi.WpsServerListener;
 import org.orbisgis.orbiswps.serviceapi.*;
 import org.orbisgis.orbiswps.service.controller.process.ProcessManager;
@@ -50,6 +52,7 @@ import org.orbisgis.orbiswps.service.execution.ProcessWorker;
 import org.orbisgis.orbiswps.service.model.JaxbContainer;
 import org.orbisgis.orbiswps.service.utils.WpsServerProperties_1_0_0;
 import org.orbisgis.orbiswps.service.utils.WpsServerProperties_2_0;
+import org.orbisgis.orbiswps.serviceapi.process.ProcessMetadata;
 import org.osgi.service.component.annotations.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
