+++ service/src/main/java/org/orbisgis/orbiswps/service/controller/parser/LiteralDataParser.java
@@ -48,7 +48,8 @@ import org.orbisgis.orbiswps.groovyapi.attributes.InputAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.LiteralDataAttribute;
 import org.orbisgis.orbiswps.service.controller.utils.ObjectAnnotationConverter;
 import org.orbisgis.orbiswps.service.model.DataType;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.parser.Parser;
 
 import javax.xml.bind.JAXBElement;
 import java.lang.reflect.Field;
