+++ service/src/main/java/org/orbisgis/orbiswps/service/controller/parser/PasswordParser.java
@@ -46,9 +46,10 @@ import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.InputAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.PasswordAttribute;
 import org.orbisgis.orbiswps.service.controller.utils.ObjectAnnotationConverter;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 import org.orbisgis.orbiswps.service.model.ObjectFactory;
 import org.orbisgis.orbiswps.service.model.Password;
+import org.orbisgis.orbiswps.serviceapi.parser.Parser;
 
 import javax.xml.bind.JAXBElement;
 import java.lang.reflect.Field;
