+++ service/src/main/java/org/orbisgis/orbiswps/service/controller/process/ProcessManager.java
@@ -44,16 +44,17 @@ import net.opengis.ows._2.CodeType;
 import net.opengis.ows._2.MetadataType;
 import net.opengis.wps._2_0.*;
 import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 import org.orbisgis.orbiswps.serviceapi.WpsServer;
 import org.orbisgis.orbiswps.service.controller.parser.ParserController;
 import org.orbisgis.orbiswps.service.controller.utils.CancelClosure;
 import org.orbisgis.orbiswps.service.controller.utils.WpsSql;
 import org.orbisgis.orbiswps.service.model.*;
 import org.orbisgis.orbiswps.service.model.Enumeration;
-import org.orbisgis.orbiswps.serviceapi.ProcessIdentifier;
-import org.orbisgis.orbiswps.serviceapi.ProcessMetadata;
+import org.orbisgis.orbiswps.serviceapi.process.ProcessIdentifier;
+import org.orbisgis.orbiswps.serviceapi.process.ProcessMetadata;
 import org.orbisgis.orbiswps.service.utils.ProgressMonitor;
-import org.orbisgis.orbiswps.serviceapi.ProcessMetadata.DBMS_TYPE;
+import org.orbisgis.orbiswps.serviceapi.process.ProcessMetadata.DBMS_TYPE;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xnap.commons.i18n.I18n;
