+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/parser/JDBCTableParserTest.java
@@ -51,7 +51,7 @@ import org.orbisgis.orbiswps.groovyapi.attributes.JDBCTableAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.OutputAttribute;
 import org.orbisgis.orbiswps.service.model.DataType;
 import org.orbisgis.orbiswps.service.model.JDBCTable;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 
 import java.lang.reflect.Field;
 import java.math.BigInteger;
