+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/parser/LiteralDataParserTest.java
@@ -51,7 +51,7 @@ import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.InputAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.LiteralDataAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.OutputAttribute;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 
 import java.lang.reflect.Field;
 import java.math.BigInteger;
