+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/parser/PasswordParserTest.java
@@ -49,7 +49,7 @@ import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.InputAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.OutputAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.PasswordAttribute;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 import org.orbisgis.orbiswps.service.model.Password;
 
 import java.lang.reflect.Field;
