+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/parser/ProcessParserTest.java
@@ -46,7 +46,7 @@ import org.junit.Before;
 import org.junit.Test;
 import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
 import org.orbisgis.orbiswps.groovyapi.attributes.ProcessAttribute;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 
 import java.lang.reflect.Method;
 import java.net.URI;
