+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/utils/DescriptionTypeConvertTest.java
@@ -47,7 +47,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.orbisgis.orbiswps.groovyapi.attributes.DescriptionTypeAttribute;
 import net.opengis.wps._2_0.DescriptionType;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
