+++ service/src/test/java/org/orbisgis/orbiswps/service/controller/utils/LiteralDataTypeConvertTest.java
@@ -45,7 +45,7 @@ import org.junit.Assert;
 import org.junit.Test;
 import org.orbisgis.orbiswps.groovyapi.attributes.*;
 import org.orbisgis.orbiswps.service.model.DataType;
-import org.orbisgis.orbiswps.service.model.MalformedScriptException;
+import org.orbisgis.orbiswps.serviceapi.data.MalformedScriptException;
 
 import java.lang.annotation.Annotation;
 import java.lang.reflect.Field;
